Support for (additional) jstree plugins
Changed the plugin to take 2 extra component attributes to allow for
loading more plugins than those required. One to define the plugin and
one with config options. These parameters are to be used as plain text
and should contain javascript object notation strings. The help of each
attribute explains more and contains examples. Tested with jquery
themes, not the default themes (yet).  The example application has been
updated to showcase a checkbox tree.